Changes per advice from review thread.
